requirements.txt
ipykernel
openai==0.28
python-dotenv
yfinance
pandas
matplotlib
langdetect
unidecode
streamlit
pymupdf
python-dotenv
fastapi
uvicorn
simple_logger